Move camera centering code from index.js to render.js
